Add option to disable input while allowing stepping and add direction
The new `readOnlyInputs` option allows to automatically disable the input while allowing the steps to change the value of the input.

There is also an added meta data to the event, which allows to know the direction of the stepping (-1 for -, 1 for +). The step count is not indicated in this direction data.