Update attributes on change
This triggers a change event every time the `val(number)` function is called on an input of a spinner.

Therefore, first of all, the programmatic way of changing a value now also triggers the event listener for change events, but now, implementations will need to deal with different kinds of events, which is the intented behavior here.